small refactor on items list
