+++ homeworks/cleandex_poketecture/lib/pages/main.page.dart
@@ -1,7 +1,7 @@
 import 'package:cleandex_poketecture/application/widgets/app_scaffold.widget.dart';
 import 'package:cleandex_poketecture/commons/app_colors.dart';
 import 'package:cleandex_poketecture/commons/interfaces.dart';
-import 'package:cleandex_poketecture/pages/partials/items_list.widget.dart';
+import 'package:cleandex_poketecture/pages/partials/items/items_list.widget.dart';
 import 'package:cleandex_poketecture/pages/partials/moves/move_list.widget.dart';
 import 'package:cleandex_poketecture/pages/partials/pokemon/pokemon_list.widget.dart';
 import 'package:flutter/material.dart';
